Decouple Model and Sql builder feature. #322
